creates test case file spec. for IM
